+++ hpe/xl420/exporter.go
@@ -88,6 +88,7 @@ type Exporter struct {
 	credProfile         string
 	biosVersion         string
 	chassisSerialNumber string
+	iloServerName       string
 	deviceMetrics       *map[string]*metrics
 }
 
@@ -177,6 +178,7 @@ func NewExporter(ctx context.Context, target, uri, profile string) (*Exporter, e
 	}
 	exp.biosVersion = resp.BiosVersion
 	exp.chassisSerialNumber = resp.SerialNumber
+	exp.iloServerName = resp.IloServerName
 
 	// vars for drive parsing
 	var (
@@ -312,7 +314,8 @@ func NewExporter(ctx context.Context, target, uri, profile string) (*Exporter, e
 	tasks = append(tasks,
 		pool.NewTask(common.Fetch(fqdn.String()+uri+"/Chassis/1/Thermal/", THERMAL, target, profile, retryClient)),
 		pool.NewTask(common.Fetch(fqdn.String()+uri+"/Chassis/1/Power/", POWER, target, profile, retryClient)),
-		pool.NewTask(common.Fetch(fqdn.String()+uri+"/Systems/1/", MEMORY_SUMMARY, target, profile, retryClient)))
+		pool.NewTask(common.Fetch(fqdn.String()+uri+"/Systems/1/", MEMORY_SUMMARY, target, profile, retryClient)),
+		pool.NewTask(common.Fetch(fqdn.String()+uri+"/Managers/1/", FIRMWARE, target, profile, retryClient)))
 
 	// DIMMs
 	for _, dimm := range dimms.Members {
@@ -470,7 +473,7 @@ func (e *Exporter) exportFirmwareMetrics(body []byte) error {
 		return fmt.Errorf("Error Unmarshalling XL420 FirmwareMetrics - " + err.Error())
 	}
 
-	(*dm)["deviceInfo"].WithLabelValues(chas.Description, e.chassisSerialNumber, chas.FirmwareVersion, e.biosVersion, XL420).Set(1.0)
+	(*dm)["deviceInfo"].WithLabelValues(e.iloServerName, e.chassisSerialNumber, chas.FirmwareVersion, e.biosVersion, XL420).Set(1.0)
 
 	return nil
 }